feat: implement onSuccess callback in SignUpForm for dashboard redirection
